Move ClaimType and ClaimTypeConstants to constants file
