+++ VAMobile/src/constants/claims.ts
@@ -1,1 +1,11 @@
+export const ClaimTypeConstants: {
+  ACTIVE: ClaimType
+  CLOSED: ClaimType
+} = {
+  ACTIVE: 'ACTIVE',
+  CLOSED: 'CLOSED',
+}
+
+export type ClaimType = 'ACTIVE' | 'CLOSED'
+
 export const MAX_NUM_PHOTOS = 10
