+++ VAMobile/src/store/slices/claimsAndAppealsSlice.test.ts
@@ -4,8 +4,8 @@ import { context, realStore, when } from 'testUtils'
 import { appeal as AppealPayload } from 'screens/BenefitsScreen/ClaimsScreen/appealData'
 import { claim as Claim } from 'screens/BenefitsScreen/ClaimsScreen/claimData'
 import { ClaimEventData, ClaimsAndAppealsGetDataMeta } from 'store/api/types'
+import { ClaimTypeConstants } from 'constants/claims'
 import { DEFAULT_PAGE_SIZE } from 'constants/common'
-import { ClaimTypeConstants } from 'screens/BenefitsScreen/ClaimsScreen/ClaimsAndAppealsListView/ClaimsAndAppealsListView'
 import { DocumentPickerResponse } from '../../screens/BenefitsScreen/BenefitsStackScreens'
 import { contentTypes } from 'store/api/api'
 import { Asset } from 'react-native-image-picker'
